LayoutTests/imported/w3c/web-platform-tests/intersection-observer/scroll-margin-nested-expected.txt


FAIL Test scroll margin intersection assert_approx_equals: intersectionRatio expected 0.2 +/- 0.001 but got 0
